move benchmark to separate file
